feat: list.js를 index.js로 변경
